+++ osf/models/registrations.py
@@ -824,6 +824,19 @@ class Registration(AbstractNode):
         if settings.SHARE_ENABLED:
             update_share(self)
 
+    def copy_registration_responses_into_schema_response(self, draft_registration):
+        """Copies registration metadata into schema responses"""
+        from osf.models.schema_response import SchemaResponse
+        schema_response = SchemaResponse.create_initial_response(
+            self.creator,
+            self,
+            self.registration_schema
+        )
+        self.registration_responses = draft_registration.registration_responses
+        schema_response.update_responses(
+            self.registration_responses
+        )
+
     class Meta:
         # custom permissions for use in the OSF Admin App
         permissions = (
